requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dill==0.3.7
filelock==3.13.1
hummingbird-ml==0.4.9
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
onnx==1.15.0
onnxconverter-common==1.14.0
packaging==23.2
protobuf==3.20.2
psutil==5.9.6
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
surrealdb-beta @ file:///Users/maxwellflitton/Documents/github/surrealDB/surrealdb.py
surrealml @ file:///Users/maxwellflitton/Documents/github/surrealDB/surrealml
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.0
typing_extensions==4.8.0
urllib3==2.1.0
